Combine 72BB0 and 73AAC to remove rodata 0
