+++ config/splat.us.dra.yaml
@@ -109,7 +109,7 @@ segments:
       - [0x41588, .rodata, 72BB0] # func_80112BB0
       - [0x41750, .rodata, 72BB0] # func_801131C4
       - [0x418C8, .rodata, 72BB0] # func_801131C4
-      - [0x418E0, .rodata, 73AAC]
+      - [0x418E0, .rodata, 72BB0]
       - [0x418F4, rodata]
       - [0x41908, rodata] # AlucardHandleDamage
       - [0x41948, rodata] # AlucardHandleDamage
@@ -180,7 +180,6 @@ segments:
       - [0x704D0, c, 704D0]
       - [0x71830, c, 71830]
       - [0x72BB0, c, 72BB0]
-      - [0x73AAC, c, 73AAC]
       - [0x75DA0, c, 75DA0]
       - [0x75F54, c, 75F54]
       - [0x7A4D0, c, 7A4D0]
